# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DZ Bus Tracker API",
    "description": "Real-time bus tracking for Algeria.\n\n* JSON uses snake_case; timestamps are RFC 3339 UTC.\n* Errors are RFC 9457 problem documents (`application/problem+json`) with a stable `code`, localized in French (default), Arabic or English via `Accept-Language`.\n* Lists use cursor pagination: `?limit=` (1–100) and `?cursor=` from `next_cursor`.\n* Cacheable GETs return a weak `ETag`; send `If-None-Match` to get `304`.\n* Rate limits are reported in `RateLimit-*` headers; `429` carries `Retry-After`.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/api-keys": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List machine-to-machine API keys.",
        "operationId": "list_api_keys",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor from a previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (1–100, default 20).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of keys (secrets are never returned)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_ApiKeyDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create a scoped API key. The secret is returned once (audited).",
        "operationId": "create_api_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Key created; store the secret now",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiKeyDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid name, scopes or expiry",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with e-mail and password.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials (unknown account, wrong password, locked or disabled account)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create a passenger account and sign it in.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created and signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "409": {
            "description": "E-mail already used",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "Too many attempts",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Request an upload URL. The file is then sent directly to storage with the returned request\n(type and exact size are signed) and attached to a resource by id before `expires_at`.\nUploads belong to a signed-in user: API keys cannot request them.",
        "operationId": "create_upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Pending upload with its presigned request (not cacheable)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Purpose not allowed for the caller, or an API key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Type not accepted or size out of range for the purpose",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "503": {
            "description": "File storage is not available (`storage_unavailable`)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_by": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string",
            "description": "Public identifier (first part of the key)."
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "description": "Result of registration or login.",
        "required": [
          "user",
          "tokens"
        ],
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/TokenPairDto"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
...
      "CreateApiKeyRequest": {
        "type": "object",
        "required": [
          "name",
          "scopes"
        ],
        "properties": {
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Optional expiry (at most two years ahead)."
          },
          "name": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Permission codes granted to the key (e.g. `tracking:read`, `notification:send`)."
          }
        },
        "additionalProperties": false
      },
...
      "CreateUploadRequest": {
        "type": "object",
        "description": "Declares a file to upload. Its type and exact size are signed into the upload URL.",
        "required": [
          "purpose",
          "content_type",
          "size_bytes"
        ],
        "properties": {
          "content_type": {
            "type": "string",
            "description": "Media type of the file, e.g. `image/jpeg` (accepted types depend on the purpose).",
            "example": "image/jpeg"
          },
          "purpose": {
            "$ref": "#/components/schemas/UploadPurposeDto"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Exact size of the file in bytes.",
            "example": 524288,
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "CreatedApiKeyDto": {
        "type": "object",
        "description": "A new key. `secret` is shown only in this response: store it securely.",
        "required": [
          "api_key",
          "secret"
        ],
        "properties": {
          "api_key": {
            "$ref": "#/components/schemas/ApiKeyDto"
          },
          "secret": {
            "type": "string"
          }
        }
      },
...
      "Language": {
        "type": "string",
        "description": "Preferred language for messages and notifications.",
        "enum": [
          "fr",
          "ar",
          "en"
        ]
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "additionalProperties": false
      },
...
      "Paginated_ApiKeyDto": {
        "type": "object",
        "description": "A page of results. Pass `next_cursor` as `cursor` to fetch the next page.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "prefix",
                "scopes",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "created_by": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                },
                "expires_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "last_used_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "name": {
                  "type": "string"
                },
                "prefix": {
                  "type": "string",
                  "description": "Public identifier (first part of the key)."
                },
                "revoked_at": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "date-time"
                },
                "scopes": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ProblemDocument": {
        "type": "object",
        "description": "RFC 9457 problem document returned by every error response.",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code."
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            }
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request path that produced the problem."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "URI identifying the problem type (`urn:dzbus:problem:<code>`)."
          }
        }
      },
      "ProblemFieldError": {
        "type": "object",
        "description": "A field error in a problem document.",
        "required": [
          "field",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine code (`required`, `invalid_email`, `too_long`, ...)."
          },
          "field": {
            "type": "string",
            "description": "Path of the offending field (`email`, `scopes[2]`)."
          },
          "message": {
            "type": "string",
            "description": "Localized human-readable message."
          },
          "params": {
            "type": [
              "object",
              "null"
            ],
            "description": "Parameters of the rule (e.g. `{\"max\": 150}`)."
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "description": "Creates a passenger account.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "amina@example.dz"
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "language": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Language"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "At least 12 characters; common and personal passwords are refused."
          },
          "phone_number": {
            "type": [
              "string",
              "null"
            ],
            "description": "Algerian mobile number (`05…`, `06…`, `07…`, `+213…`)."
          }
        },
        "additionalProperties": false
      },
...
      "TokenPairDto": {
        "type": "object",
        "description": "Bearer token pair. Clients must serialise refresh calls: presenting a spent refresh token\nrevokes the whole session.",
        "required": [
          "token_type",
          "access_token",
          "expires_in",
          "access_token_expires_at",
          "refresh_token",
          "refresh_token_expires_at",
          "session_id"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "access_token_expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "description": "Seconds until the access token expires."
          },
          "refresh_token": {
            "type": "string"
          },
          "refresh_token_expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "session_id": {
            "type": "string",
            "format": "uuid"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          }
        }
      },
...
      "UploadDto": {
        "type": "object",
        "description": "A pending upload. Upload the file with `upload` before `expires_at`, then pass `id` to the\nendpoint that attaches it (e.g. `PUT /me/avatar`).",
        "required": [
          "id",
          "purpose",
          "content_type",
          "size_bytes",
          "upload",
          "expires_at"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "purpose": {
            "$ref": "#/components/schemas/UploadPurposeDto"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "upload": {
            "$ref": "#/components/schemas/PresignedRequestDto"
          }
        }
      },
      "UploadPurposeDto": {
        "type": "string",
        "description": "What an upload will be attached to.",
        "enum": [
          "avatar",
          "driver_id_card",
          "driver_license",
          "bus_photo",
          "stop_photo"
        ]
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Machine-to-machine key (`dzk_…`) with explicit scopes."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from /auth/login or /auth/refresh (EdDSA)."
      }
    }
  },
...
}